Major update
Now using esprima to parse constraints instead of regexes (seriously,
what was I thinking? xD)

This makes it a lot more reliable, although I still do need to add a
bunch of test cases just to make sure.

Also, with this update comes a new syntax and the capability for each
element/property in an array/object to have its own type.